Unify formula gen and formula tests.
